Optimal FIFO Route grouping

+++ DataStructures/Intermediate/Entities/StopEvent.h
@@ -39,7 +39,7 @@ public:
 
     friend std::ostream& operator<<(std::ostream& out, const StopEvent& s)
     {
-        return out << "StopEvent{" << s.stopId << ", " << s.arrivalTime << ", " << s.departureTime << "}";
+        return out << "StopEvent{" << size_t(s.stopId) << ", " << s.arrivalTime << ", " << s.departureTime << "}";
     }
 
     inline void serialize(IO::Serialization& serialize) const noexcept
